20140120-commit

+++ FISP_git/src/com/synesoft/ftzmis/app/controller/FTZ210310Controller.java
@@ -31,7 +31,7 @@ import com.synesoft.ftzmis.domain.model.FtzOffTxnDtl;
 import com.synesoft.ftzmis.domain.service.FTZ210310Service;
 
 /**
- * @author hb_huang
+ * @author hb_huang	
  * @system FTZMIS(远期结售汇)
  * @date 2014-1-3上午10:01:59
  */
@@ -97,6 +97,8 @@ public class FTZ210310Controller {
 	public String addDtlInit(Model model, FTZ210310Form form) {
 		FtzOffMsgCtl ftzOffMsgCtl = new FtzOffMsgCtl();
 		ftzOffMsgCtl.setMsgStatus(CommonConst.FTZ_MSG_STATUS_INPUTING);
+		ftzOffMsgCtl.setBranchId(ContextConst.getCurrentUser().getLoginorg());
+		ftzOffMsgCtl.setWorkDate(DateUtil.getNowOutputDate());
 		form.setFtzOffMsgCtl(ftzOffMsgCtl);
 		form.setActionFlag("addMsg");
 		model.addAttribute("pageUrl", "/FTZ210310/AddDtlInit");
@@ -152,7 +154,7 @@ public class FTZ210310Controller {
 		if (i < 0) {
 			model.addAttribute(ResultMessages.error().add("e.sysrunner.0006"));
 		} else {
-			model.addAttribute(ResultMessages.success().add("i.sm.0001"));
+			model.addAttribute(ResultMessages.success().add("ftzmis.Add.Msg.Ctl.Success"));
 		}
 		form.getFtzOffMsgCtl().setWorkDate(
 				DateUtil.getFormatDateAddSprit(form.getFtzOffMsgCtl()
@@ -328,7 +330,7 @@ public class FTZ210310Controller {
 		if (i < 0) {
 			model.addAttribute(ResultMessages.error().add("e.sysrunner.0006"));
 		} else {
-			model.addAttribute(ResultMessages.success().add("i.sm.0001"));
+			model.addAttribute(ResultMessages.success().add("ftzmis.Add.Txn.Dtl.Success"));
 			model.addAttribute("updFlag", "1");
 		}
 		form.getFtzOffTxnDtl().setSubmitDate(
@@ -493,7 +495,7 @@ public class FTZ210310Controller {
 			model.addAttribute("pageUrl", "/FTZ210310/UpdDtlInit");
 			return "ftzmis/FTZ210310_Input_Dtl";
 		} else {
-			model.addAttribute(ResultMessages.success().add("i.dp.mpp.0002"));
+			model.addAttribute(ResultMessages.success().add("ftzmis.Update.Msg.Ctl.Success"));
 		}
 		
 		form.getFtzOffMsgCtl().setWorkDate(
@@ -639,7 +641,7 @@ public class FTZ210310Controller {
 			model.addAttribute(ResultMessages.error().add(
 			"e.ftzmis.210101.0026"));
 		} else {
-			model.addAttribute(ResultMessages.success().add("i.dp.mpp.0002"));
+			model.addAttribute(ResultMessages.success().add("ftzmis.Update.Txn.Dtl.Success"));
 			model.addAttribute("updFlag", "1");
 		}
 		form.getFtzOffTxnDtl().setSubmitDate(
@@ -661,14 +663,42 @@ public class FTZ210310Controller {
 	public String delDtl(Model model, FTZ210310Form form) {
 		FtzOffMsgCtl del_FtzOffMsgCtl = new FtzOffMsgCtl();
 		del_FtzOffMsgCtl.setMsgId(form.getSelected_msgId());
-
+		
+		FtzOffTxnDtl query_FtzOffTxnDtl = new FtzOffTxnDtl();
+		query_FtzOffTxnDtl.setMsgId(form.getSelected_msgId());
+		List<FtzOffTxnDtl> ftzOffTxnDtls = ftz210310Service.
+				queryFtzOffTxnDtlList(query_FtzOffTxnDtl);
+		if (null != ftzOffTxnDtls) {
+			for(FtzOffTxnDtl ftzOffTxnDtl : ftzOffTxnDtls) {
+				if (CommonConst.FTZ_MSG_STATUS_AUTH_SUCC.equals(ftzOffTxnDtl
+						.getChkStatus())) {
+					model.addAttribute(ResultMessages.error().add(
+					"e.ftzmis.210101.0035"));
+					form.setSelected_msgId("");
+					return "forward:/FTZ210310/AddQry";
+				}
+			}
+		}
+		
+		//查询数据
+		FtzOffMsgCtl result_FtzOffMsgCtl = ftz210310Service.queryFtzOffMsgCtl(del_FtzOffMsgCtl);
+		if (!CommonConst.FTZ_MSG_STATUS_INPUTING.equals(result_FtzOffMsgCtl.getMsgStatus()) 
+				&& !CommonConst.FTZ_MSG_STATUS_INPUT_COMPLETED.equals(result_FtzOffMsgCtl.getMsgStatus())
+				&& !CommonConst.FTZ_MSG_STATUS_AUTH_FAIL.equals(
+						result_FtzOffMsgCtl.getMsgStatus())) {
+			model.addAttribute(ResultMessages.error().add(
+			"e.ftzmis.210101.0036"));
+			form.setSelected_msgId("");
+			return "forward:/FTZ210310/AddQry";
+		}
+		
 		int i = ftz210310Service.deleteFtzOffMsgCtl(del_FtzOffMsgCtl);
 
 		if (i < 0) {
 			model.addAttribute(ResultMessages.error().add("e.sysrunner.0002"));
 			form.setSelected_msgId("");
 		} else {
-			model.addAttribute(ResultMessages.success().add("i.dp.0003"));
+			model.addAttribute(ResultMessages.success().add("ftzmis.Delete.Msg.Ctl.Success"));
 			form.setSelected_msgId("");
 			return "forward:/FTZ210310/AddQry";
 		}
@@ -688,7 +718,7 @@ public class FTZ210310Controller {
 			model.addAttribute(ResultMessages.error().add("e.sysrunner.0002"));
 			form.setSelected_seqNo("");
 		} else {
-			model.addAttribute(ResultMessages.success().add("i.dp.0003"));
+			model.addAttribute(ResultMessages.success().add("ftzmis.Delete.Txn.Dtl.Success"));
 			form.setSelected_seqNo("");
 			return "forward:/FTZ210310/UpdDtlInit";
 		}
@@ -721,7 +751,7 @@ public class FTZ210310Controller {
 			return "forward:/FTZ210310/AddQry";
 		}
 		
-		model.addAttribute(ResultMessages.success().add("i.ftzmis.210101.0001"));
+		model.addAttribute(ResultMessages.success().add("ftzmis.Submit.Msg.Ctl.Success"));
 		return "forward:/FTZ210310/AddQry";
 	}
